separate functions for compressed loading and saving
